pyproject.toml
[project]
name = "mikrotik-proxy-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "loguru>=0.7.3",
    "pydantic-settings>=2.8.1",
    "pyyaml>=6.0.2",
    "routeros-api>=0.21.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.9",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # bugbear
    "UP",  # pyupgrade
    "SIM", # simplify
    "RUF", # ruff-specific
    "A",   # builtins shadowing
]
ignore = [
    "E501",  # line length is handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]  # asserts in tests are fine
